+++ frontend/src/worksheets/TableListOfWorkSheets.js
@@ -8,7 +8,7 @@ import useCurrentUser from '../hooks/useCurrentUser'
 import CloseButton from '../specialButtons/CloseButton'
 import { closeWorkSheet, finalizeWorkSheet } from '../api/WorkSheetAPI'
 import LoadingScreen from '../LoadingScreen'
-import workSheetPDF from './WorkSheetPDF'
+import workSheetPDF from './workSheetPDF'
 import PdfButton from '../specialButtons/PdfButton'
 import FinalizeButton from '../specialButtons/FinalizeButton'
 import CalendarDropDown from '../CalendarDropDown'
