+++ frontend/src/worksheets/workSheetPDF.js
@@ -6,18 +6,7 @@ import {
   assetSettlement,
   typeOfPayment,
 } from '../TranslationForWorkSheet'
-import SignaturePad from 'signature_pad'
-
-function renderSvg(signatureData) {
-  const canvasSignature = document.createElement('canvas')
-  canvasSignature.width = 400
-  canvasSignature.height = 300
-  const signaturePad = new SignaturePad(canvasSignature)
-  signaturePad.fromData(JSON.parse(param))
-  const svgSignature = signaturePad.toDataURL('image/svg+xml')
-  const data = svgSignature.split(',')[1]
-  return atob(data)
-}
+import renderSvg from './renderSvg'
 
 const acknowledge =
   'A munkavégzést igazoló aláírásával a fent megjelölt munka teljesítését elismeri, az üzemelő rendszert átveszi.'
